requirements.txt
python-telegram-bot
python-mk-livestatus
